code review corrections
