+++ src/main/java/com/artipie/front/auth/Credentials.java
@@ -4,7 +4,6 @@
  */
 package com.artipie.front.auth;
 
-import com.amihaiemil.eoyaml.YamlMapping;
 import java.util.Optional;
 
 /**
@@ -19,31 +18,4 @@ public interface Credentials {
      */
     Optional<User> user(String name);
 
-    /**
-     * Yaml credentials parser.
-     * @since 1.0
-     */
-    final class FromYaml implements Credentials {
-
-        /**
-         * Yaml source.
-         */
-        private final YamlMapping mapping;
-
-        /**
-         * New yaml credentials.
-         * @param mapping Yaml
-         */
-        public FromYaml(final YamlMapping mapping) {
-            this.mapping = mapping;
-        }
-
-        @Override
-        public Optional<User> user(final String name) {
-            return Optional.ofNullable(this.mapping.yamlMapping("credentials"))
-                .flatMap(cred -> Optional.ofNullable(cred.yamlMapping(name)))
-                .map(yaml -> new User.FromYaml(yaml, name));
-        }
-
-    }
 }
