+++ src/main/java/com/artipie/front/auth/User.java
@@ -4,14 +4,8 @@
  */
 package com.artipie.front.auth;
 
-import com.amihaiemil.eoyaml.YamlMapping;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
-import org.apache.commons.codec.digest.DigestUtils;
 
 /**
  * User.
@@ -49,92 +43,4 @@ public interface User {
      */
     Optional<String> email();
 
-    /**
-     * Yaml user item.
-     * @since 1.0
-     */
-    final class FromYaml implements User {
-
-        /**
-         * Yaml source.
-         */
-        private final YamlMapping mapping;
-
-        /**
-         * User name (id).
-         */
-        private final String name;
-
-        /**
-         * New user.
-         * @param mapping Yaml
-         * @param name User name
-         */
-        public FromYaml(final YamlMapping mapping, final String name) {
-            this.mapping = mapping;
-            this.name = name;
-        }
-
-        @Override
-        public boolean validatePassword(final String pass) {
-            final var config = this.mapping.string("pass");
-            if (config == null) {
-                throw new IllegalStateException(
-                    "invalid credentials configuration: `pass` field not found"
-                );
-            }
-            final var type = this.mapping.string("type");
-            final boolean res;
-            if (type == null) {
-                res = validateStringPass(config, pass);
-            } else {
-                res = validateStringPass(String.join(":", type, config), pass);
-            }
-            return res;
-        }
-
-        @Override
-        public String uid() {
-            return this.name;
-        }
-
-        @Override
-        public Set<? extends String> groups() {
-            return Optional.ofNullable(this.mapping.yamlSequence("groups"))
-                .map(seq -> StreamSupport.stream(seq.spliterator(), false))
-                .orElse(Stream.empty())
-                .map(node -> node.asScalar().value())
-                .collect(Collectors.toSet());
-        }
-
-        @Override
-        public Optional<String> email() {
-            return Optional.ofNullable(this.mapping.string("email"));
-        }
-
-        /**
-         * Validate password string.
-         * @param config Passowrd string config
-         * @param pass Actual password
-         * @return True if password is valid
-         * @checkstyle ReturnCountCheck (30 lines)
-         */
-        @SuppressWarnings("PMD.OnlyOneReturn")
-        private static boolean validateStringPass(final String config, final String pass) {
-            final var parts = config.split(":");
-            switch (parts[0]) {
-                case "plain":
-                    return Objects.equals(parts[1], pass);
-                case "sha256":
-                    return Objects.equals(parts[1], DigestUtils.sha256Hex(pass));
-                default:
-                    throw new IllegalStateException(
-                        String.format(
-                            "invalid credentials configuration: `pass` type `%s` is not supported",
-                                parts[0]
-                        )
-                    );
-            }
-        }
-    }
 }
